refactor: view type 공용 타입으로 변경 (#ATR-576)
