+++ apps/service/src/entities/user-article/model/type/index.ts
@@ -1,4 +1,3 @@
 export * from './article'
-export * from './viewType'
 export * from './sortType'
 export * from './newsletter'
