+++ apps/service/src/features/filter-article/model/useViewType.ts
@@ -1,7 +1,7 @@
 'use client'
 
 import { useCallback, useState } from 'react'
-import { ViewType } from '@/entities/user-article'
+import type { ViewType } from '@/shared/type'
 
 export default function useViewType() {
   const [viewType, setView] = useState<ViewType>('gallery')
