+++ apps/service/src/shared/type/viewType.ts
